refactor: address code review feedback
- Extract detectWidget() to src/utils/widget.ts to avoid
  cross-boundary import from CF Workers module into Node script
- Randomize User-Agent per member instead of per cron run
- Document known limitation: prev/next links are not checked
  against a specific member slug
